requirements.txt
pandas
polars
python-dotenv
requests
psycopg2-binary
scikit-learn
numpy